requirements.txt
pyClamd>=0.3.17
django-password-policies>=0.4.1
django-axes>=1.3.8